Added balancer pool
