+++ src/helpers/price-updates.ts
@@ -18,12 +18,31 @@ export function savePriceToHistory(oracleAsset: PriceOracleAsset, event: ethereu
   priceHistoryItem.save();
 }
 
-export function updateDependentAssets(
-  dependentAssets: string[],
-  event: ethereum.Event): void {
+// Method called for external pool updates Uniswap / balancer etc
+export function updateAssetPriceFromAaveOracle(event: ethereum.Event): void {
+  let assetAddress = event.address;
+  let priceOracle = getOrInitPriceOracle();
+  let priceOracleAsset = getPriceOracleAsset(assetAddress.toHexString());
+  let proxyPriceProvider = AaveOracle.bind(priceOracle.proxyPriceProvider as Address);
+
+  let assetPriceCall = proxyPriceProvider.try_getAssetPrice(assetAddress);
+  if (!assetPriceCall.reverted) {
+    priceOracleAsset.priceInEth = assetPriceCall.value;
+    priceOracleAsset.save();
+
+    // save price to history
+    savePriceToHistory(priceOracleAsset, event);
+  } else {
+    log.error('Error in getting price from uniswap price feed for asset: {}', [
+      assetAddress.toHexString(),
+    ]);
+  }
+}
+
+export function updateDependentAssets(dependentAssets: string[], event: ethereum.Event): void {
   let proxyPriceProviderAddress = getOrInitPriceOracle().proxyPriceProvider;
   let proxyPriceProvider = AaveOracle.bind(proxyPriceProviderAddress as Address);
-  
+
   // update dependent assets price
   for (let i = 0; i < dependentAssets.length; i += 1) {
     let dependentAsset = dependentAssets[i];
